move tests to test folder and refactor quit()
