+++ lib/redis.js
@@ -31,14 +31,14 @@ class RedisClient {
 	_attachRedisEvent() {
 		this._redis_client.on("end", (err) => {
 			// Exit application when Redis session is ended.
-			process.exit(0, false);
+			process.exit(0);
 		});
 
 		this._redis_client.on("error", (err) => {
 			console.log(colors.red("(error) %s"), err.message);
 
 			// Return non-zero value for error.
-			this._exit(-1);
+			process.exit(-1);
 		});
 	}
 
@@ -96,7 +96,7 @@ class RedisClient {
 						let CMD = commands.shift().toLowerCase();
 						//`exit` and `clear` are not true commands, just part of REPL
 						if (CMD === 'exit') {
-							this._exit();
+							this._redis_client.quit();
 						} else if (CMD === 'clear') {
 							console.log('\x1b[0f'); /* ANSI clear screen code */
 							readline.cursorTo(process.stdout, 0, 0);
@@ -112,18 +112,15 @@ class RedisClient {
 					rl.prompt();
 				}
 			}).on('close', () => {
+				// trigger when `SIGINT` received
 				console.log('\nAbort!');
-				this._exit();
+				this._redis_client.quit();
 			});
 		});
 	}
 
-	_exit(code = 0, exit = true) {
-		if (exit) {
-			return this._redis_client.quitAsync().then(() => process.exit(code));
-		} else {
-			process.exit(code);
-		}
+	get client() {
+		return this._redis_client;
 	}
 }
